requirements.txt
aiohttp==3.9.1
boto3==1.34.14
google-api-python-client==2.108.0
google-auth==2.20.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
openpyxl==3.1.2
pandas==2.1.4
pdf2image==1.16.3
Pillow==10.1.0
pinecone-client==3.0.2
python-docx==1.0.1
python-dotenv==1.0.0
python-pptx==0.6.22
tiktoken==0.5.2
tqdm==4.66.1
